content/canvas/test/test_size.attributes.parse.badsuffix.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.parse.badsuffix</title>
<!-- Testing: Parsing of non-negative integers -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100foo" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

todo(canvas.width == 100, "canvas.width == 100");

SimpleTest.finish();

});
</script>
